Change layout of PC's stat pools
I think this makes the relation between a pools current value and size
more obvious.  I found when my friends and I were playing and making
characters, that we constantly forgot about the "pool" value. I had to
explain the relationship between the text boxes as "fractions".  I
thought this might be a better way to visually show the two values. I
noticed that the NPC sheet already displays health this way.

After making that change it looked weird having Edge take up so much
space so I shrunk it down to not be as large.

Signed-off-by: Taylor Silva <taylorsilva@outlook.com>